feat: SSG for dynamic routes
This commit addresses the problem where dynamic routes were not being
statically rendered. While a dynamic route can't be statically rendered,
it can leverage ISR when `generateStaticParams` is used. This is exactly
what we want.

I tried this with the dynamic Route Handler but it didn't work as it
failed with a "dynamic server error". See more here:
https://nextjs.org/docs/messages/dynamic-server-error.