feat: add Draft Mode proxy
Forwarding the request in middleware with the `__prerender_bypass`
cookie doesn't work when deployed on Vercel. This is likely because the
decision to bypass the edge cache happens before the middleware is
invoked, so attempting to forward the cookie in middleware might be too
late. Our approach instead is to add the Draft Mode cookie in a proxy.